requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
Pillow